Refactored Setup and moved device setup for local and device farms to separate files.
